Fix an issue where macros were not applied compiling full modules (#26)
* Fix an issue where macros were not applied compiling full modules

In a previous PR, work was done to fully compile Basilisp namespaces in a single
file down into a corresponding Python module which would be usable as a standard
Python module. This caused trouble applying macros during the compilation phase,
because it meant that the entire namespace was compiled and executed at once,
rather than piecemeal. Without the runtime updating with each addition of a
function or definition to a namespace, macros did not exist and no code was
generated.

This PR adds a hack on top of the compiler to run incremental compilation of
Basilisp code in parallel with the standard module generation process to allow
the runtime to be built up as the module is being compiled (and therefore allow
macros to exist).

* Fix type check error