Allow users to disable emitting Python AST strings (#356)
Emitting the generated Python AST to strings and then to stdout or `*generated-python*` Var adds a bit of time to startup, which is a fine price to pay during development, but would be more onerous to normal users. This PR creates an environment var `BASILISP_EMIT_GENERATED_PYTHON` which can be set to disable the emission of generated Python. Eventually, the default value will switch to `false` (when development settles down), but for now having `true` set as the default value is tremendously useful.