Add the ILookup interface (#410)
This commit adds the `ILookup` interface, which is inserted into the `IAssociative` hierarchy. `ILookup` adds `ILookup.val_at`, which returns only values from associative objects and `IAssociative.entry` is amended to return an `IMapEntry`. 

For #409 we need to apply the `ILookup` interface to `ReaderConditional` types. Before, Basilisp only had the `IAssociative.entry` which was always meant to return an `IMapEntry` (rather than merely a value, as it was doing).