Add JSON encoder and decoder (#485)
Add a JSON encoder and decoder based on Python's `json` module.

This is obviously not going to be the fastest encoder or decoder in the planet, but it will be builtin so folks don't have to fetch another library to do the job. It's made worse by the fact that Python's builtin decoder does not include an `array_hook` option. We could probably make a fairly efficient decoder with that and the existing `object_hook`, but since we have the completely decode the entire object to Python objects and then convert, it's probably pretty slow.