Support artificially abstract base classes (#565)
Add support for "artificially abstract" base classes. The Python ecosystem is much less strict with its use of `abc.ABC` to interfaces than Java (which has a native `interface` construct), so even in where a type may be _in practice_ an interface or ABC, the compiler would not permit you to declare such types as supertypes because they do not inherit from `abc.ABC`. In these cases, users can mark the type as  abstract with the `:abstract` metadata key to force the compiler to accept that it is abstract.

In this case, the compiler will attempt to verify that any extra methods present on the `deftype`, `defrecord`, or `reify` are present on any one of the artificially abstract base classes. This will allow users to override existing class members, but not introduce new ones without a true interface.