Lock PyTest to 6.2.5; allow Black to update freely
For now it appears that supporting PyTest 7.0 will take a bit more work than
I have time for, so I'm just going to lock it to 6.2.5 and file a separate
ticket to address support.

Black is no longer "beta", so we can open up the version lock to "*" and let
it float with other dev dependencies.