Delete generated utility functions from REPL namespace after execution (#674)
Fixes #675 

In order to fetch the expression from an expression, the Basilisp REPL generates a simple function in the current REPL module which returns the expression. Afterwards, the function definition hangs around indefinitely even though the function will never be used or called again. This commit adds code to simply `del` that function from the module so it doesn't hang around eating up memory.