Fix MS-Windows specific failures in test suite. (#688)
Hi,

could you please review changes to fix the test suite to pass on
MS-Windows. It fixes #687.

It also enables the test suite which includes MS-Windows support to pass
in #686.

In addition to the document changes, the patch also
- Releases file/dir before removing (monkeypatch.chdir, os/close).
MS-Windows will complain about trying to remove a file that is still
being used otherwise.
- Add empty server to file: uri (i.e. file:///), since paths such as
c:\dev, will pickup c: as the server causing the tests to fail.
- Just looks for the file name in traceback texts (e.g.
*test_testrunner.lpy instead of /*/test_testrunner.lpy), otherwise it
will fail because the output string will be c:\*\test_testrunner.lpy on
windows.

The test suite currently fails to run because package dependencies have
moved since it was last run. Perhaps is best to merge this after #686.

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>
Co-authored-by: Chris Rink <chrisrink10@users.noreply.github.com>