`case` should only evaluate the matching clause expr (#700)
Hi,

can you please consider a patch to evaluate only the expression that
matches the clause. It fixes #699 .

The issue was that `~(apply hash-map pairs)` evaluated all the clauses
expressions.

With this update, the clauses are turned into `cond` clauses, taking
extra case to quote any symbol bindings.

I tried to keep as match of the old code as possible, I could not think
of any other way of doing so. It seems as if this requires a special
form to as not to evaluate the clauses (e.g. `cond`). The drawback with
this approach is that we went from O(1) to find the right expr to return
it's value, to O(n).

Feel free to update or discard in case there is an obvious better
solution.

I've included a test to confirm the same.

Thanks,

(the corresponding def in clojure-1.11 is [quite
involved](https://github.com/clojure/clojure/blob/ce55092f2b2f5481d25cff6205470c1335760ef6/src/clj/clojure/core.clj#L6748))

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>