Store Poetry binary in the cache (#835)
There was an issue in #827 where subsequent runs of the minimum
dependencies job would fail due to a missing Poetry binary. I believe
the cause of that was a missing cache path. The cache paths were copied
from the other `run-tests` job which was also wrong, but which did not
ever try to execute `poetry` binary directly so the missing binary issue
was never discovered.