Fix where `(str seq)` was printing seq string items without `"` (#892)
Hi,

could you please consider patch to have the `str` function on seq string
items surrounding them with quotation marks. It fixes #891

The fix simply resets the `human_readable` kw argument for inner
elements to False for seqs and maps.

I have included a `test_str` test in `lrepr_test.py` to test the same.
This is based on `test_lstr` above it.

It appears to me as if `test_lstr` is inaccurately named. Despite being
labeled `lstr`, it employes `pr-str` which differs from invoking `lstr`
directly with its arguments, Therefore, I find naming it `test_lstr` a
little misleading; it rather seems as to be only capturing the output of
`pr`. I could be wrong though 😅

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>