*print-namespace-maps* support (#883)
Hi,

could you please review draft support for the *print-namespace-maps*
dynamic variable. It resolves #882.

I implemented the support at the lowest `obj.py` level, though this
required to bring in the `keyword.py` and `symbol.py` modules resulting
to cyclic dependencies. Not sure how else to implement it, so any other
suggestions are most welcome.

I've also blindly followed the example of *print-dup* on how I ended
setup the variable in the cli/runtime/object .py files.

In addition, I've used a set to compare the printed output of the map
result with multiple k/v, exhaustively covering the possible orderings,
is there perhaps a better simpler way to do this? This currently is
prune to user errors.

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>