ensure nREPL server can exit even when client connections are open (#1003)
Hi,

can you please review patch to let nREPL server to exit even when client
connections are still open. It addresses #1002.

This is simply done by setting the server's `daemon-threads` field to
true.

Additionally, I've updated the `:server*` option accepted by
`server-make!` from an `atom` to a `promise`. This change aligns with
the original intent of providing a server reference when it becomes
available. Note that this is a breaking change as it alters the API.
Given the current development stage and no known users of this API, I
believe it's safe to include this change now. Let me know if you'd
prefer this in a separate PR or think otherwise.

I've added a test for the same.

Thanks

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>