Changed eager read-all to lazy read-seq #999 (#1000)
Fixes #999 

Looks like eagerly reading forms breaks macros.

This replaces `read-all` with the lazy `read-seq` function. 

This being lazy means the user now needs to be aware of when the seq is
realized to avoid trying to do so after the reader is closed. This is
something I was hoping to avoid that but it might not be possible.

---------

Co-authored-by: Chris Rink <chrisrink10@users.noreply.github.com>