Align `basilisp.core/deref` with Clojure to use timeout in ms (#1008)
Hi,

could you please consider patch to update the `basilisp.core/deref`
timeout argument to us ms instead of secs. It addresses #1007

This is to align with the same in `clojure.core/deref`.

I have add a `promise` basilisp test and updated an existing `future`
one. The tests check that the timeout for an unfulfilled promise returns
within a range of 100ms. I believe this range provides sufficient buffer
to avoid interference from parallel processing on the host.

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>