conform to spec on nrepl's info op on :file and :column values (#1068)
Hi,

could you please review patch to align the nrepl server's `info` op with
the `cider-nrepl` spec. It fixes #1066.

According to the spec, the `:file` return value should be a URI when it
refers to an absolute path. It also specifies `:column` as a return key,
which was previously missing.

`cider-nrepl` `info` spec:
https://docs.cider.mx/cider-nrepl/nrepl-api/ops.html#info

The  nrepl-server's `info` test has been updated accordingly.

Thanks

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>