Do not attempt to resolve symbols in data reader tags (#1132)
Fixes #1129 

For anyone who finds this PR (perhaps triaging a bug introduced by it in
the future): I tried to ascertain exactly what Clojure JVM does in this
situation and it is quite unclear in the reader code and not (at the
time of writing) specified on any of the official documentation.
Examples given there seem to indicate that fully qualified tags be given
in `data_readers.cljc` and thus symbol resolution should not be
performed on them, but they don't say explicitly. CLJS `#js []` literals
work as they do now in this PR so I'm going to assume this is the
intended behavior.