Update alter-var-root to return new val, updated docs (#1171)
Hi,

can you please review change to have have `alter-var-root` to return the
new value. It addresses #1166.

I've also updated various part of the documentation to note the effect
of direct linking on this fn. Please feel free to adjust partially or
completely.

I've also added test cases for the `alter-var-root` for the new change
plus to demo the direct linking behavior.

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>
Co-authored-by: Chris Rink <chrisrink10@users.noreply.github.com>