return nil from `import` (#1177)
Hi,

can you please review patch to have `import` to return `nil`. It
addresses #1174.

I haven't included a test for this case, nor mentioned it in the
docstring, implying the return value is unspecified. I'm not sure this
is something we want to fix at the API level (although I did request it
as an aesthetic addition).

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>