Adds support for :decorators in anon fns, defasync (#1189)
Hi,

could you please review patch to support the `:decorators` key in
anonymous `fn`s. It addresses #1178.

The handling of `:decorators` has been nmoved to `fn`, which also
ensures that the meta key is now considered when passed as metadata to
the `defn` name. This enables support for supporting the meta key in
`defn` derived macros such as `defasync`, as intended.

I've updated the `defn` docstring to hint that the processing is now
done by `fn` now.

I've also added tests for `fn`, `defn` and `defasync`, along with a
section in the python interop doc about it.

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>
Co-authored-by: Chris Rink <chrisrink10@users.noreply.github.com>