fix testrunner to handle relative entries in sys.path (#1205)
Hi,

could you please review patch to fix a likely regression issue where the
testrunner couldn't handle relative paths in `sys.path`. It fixes #1204.

This is likely a regression from #1176, where the update to the new
module loading system caused relative paths to never match a fully
qualified module name. The patch now expands relative paths to their
absolute form before comparing them to the module path.

This issue was breaking the vanilla `basilisp test` command, as it adds
an empty string (`""`) entry to `sys.path`, which is interpreted as
`"."`. Since this is a relative path, it wasn’t handled by the logic,
causing the project directory not to be recognized as the root of the
Basilisp test modules.

I’ve added a test that invokes `basilisp test` directly as a subprocess.
Unfortunately, I couldn’t find a way to control the pytester `sys.path`
entries with the precise control required for this test. Any counter
suggestions are most welcome.

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>