Pickup bootstrap site-packages location from sysconfig purelib path (#1211)
Hi,

could you please review patch to pickup the boostrap site-packages
location from the `sysconfig` `purelib` path. It fixes #1208.

The bootstrap options and code now operate on a single path instead of a
list, likely because `getsitepackages` previously returned a list of
paths, which is no longer relevant (unless there's another reason I'm
missing).

I've also updated the `bootstrap` command to print the file's creation
path, which I find useful for awareness.

For testing, I couldn't think of a case other than creating a virtual
environment to test the `bootstrap` within it. I implemented a minimal
solution using the `pytest_virtualenv` plugin, which provides a fixture
for this. Since creating a Basilisp environment takes 20+ seconds, I
added a pytest decorator to run it only on GitHub CI. Let me know if you
have other testing suggestions or if it's better to leave this out
altogether. I consider these type of tests as part of integration
testing (which ideally run as a separate command/job), but we don't have
that setup (yet?).

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>