Make support for single-use iterables explicit (#1235)
Hi, 

could you please consider patch to remove explicit support of single-use
iterables in sequences. It fixes #1192.

single-use iterables now throw a `TypeError`, when they are implicitly
or explicitly coerced into a sequence. A new `iterator-seq` is
introduced to explicitly convert them into a seq instead . See
https://github.com/basilisp-lang/basilisp/pull/1198#issuecomment-2810302395
for the rationale.

The `count` fn was also updated to behave the same.

There were a few places where single use iterators were converted
implicitly to a seq, which are now explicitly converted using a
`iterator-seq` or the python equivalent instead.

Tests are added for the same, as well as documentation section under
Python Iterators.

Thanks

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>