rename awhile, afor to while-async, for-async (#1250)
Hi,

can you please review patch to rename the `awhile` and `afor` macros to
`while-async` and `for-async` for better clarity. It addresses the point
about naming in #1248.

I felt placing `async` after `for`/`while` reads more naturally, but I'm
open to other suggestions if you think a better naming.

Thanks

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>