suppress pytest rewrite assertion warning (#1279)
Hi,

can you please consider patch to suppress the pytest rewrite assertion
warning when running `basilisp test`. It fixes #1252.

I haven’t added a test, as it would require creating a virtual
environment during testing, which seemed excessive for this issue.
However, I’m happy to add one if you feel it’s necessary.

Thanks

Note: this is a follow up to #1255 but it uses the `;` separator to
specify the module name, introduced by
https://github.com/pytest-dev/pytest/pull/13429 in [Pytest
8.4.0](https://docs.pytest.org/en/stable/changelog.html#pytest-8-4-0-2025-06-02).

---------

Co-authored-by: ikappaki <ikappaki@users.noreply.github.com>